v0.69 PR3 (partial Phase 3) — split video-review, noise-reduce, fade
Begins Plan G Phase 3 (ai-edit.ts split). 3 of 8 sections moved into
`packages/cli/src/commands/_shared/edit/`:

- video-review.ts (49 L) — types only (AutoFix, VideoReviewCategory,
  VideoReviewFeedback) used by ai-review.ts via Gemini analysis
- noise-reduce.ts (100 L) — executeNoiseReduce + types (FFmpeg afftdn)
- fade.ts (127 L) — executeFade + types (FFmpeg fade/afade)

ai-edit.ts: 1,589 → 1,345 lines (−244). Stays as a barrel: re-exports
from each new file so the 9 consumer paths (demo, edit-cmd, ai-edit-cli,
ai-review, ai-script-pipeline-cli, ai-animated-caption, ai,
ai-script-pipeline, ai.test) keep working unchanged.

## Remaining (v0.69 PR4 / next session)

- silence-cut (320 L) — biggest section, FFmpeg silencedetect + Gemini fallback
- jump-cut (278 L) — Whisper word-level + filler removal
- caption (236 L) — Whisper + ASS subtitles burn-in
- translate-srt (170 L) — Claude/OpenAI SRT translation
- text-overlay (290 L) — applyTextOverlays + executeTextOverlay

After all 8 split, ai-edit.ts becomes ~50-line barrel only.

## Test plan

- [x] `pnpm -r exec tsc --noEmit` — 0 errors
- [x] `pnpm lint` — 0 errors
- [x] `pnpm -F @vibeframe/cli test` — 652 passed (unchanged)
- [x] All re-exports preserve external import paths
