+++ packages/cli/src/commands/ai-edit.ts
@@ -863,221 +863,11 @@ export async function executeCaption(options: CaptionOptions): Promise<CaptionRe
   }
 }
 
-// ============================================================================
-// Noise Reduce
-// ============================================================================
-
-/** Options for {@link executeNoiseReduce}. */
-export interface NoiseReduceOptions {
-  /** Path to the input audio or video file */
-  inputPath: string;
-  /** Path for the noise-reduced output file */
-  outputPath: string;
-  /** Reduction strength preset (default: "medium") */
-  strength?: "low" | "medium" | "high";
-  /** Custom noise floor in dB (overrides strength preset) */
-  noiseFloor?: number;
-}
-
-/** Result from {@link executeNoiseReduce}. */
-export interface NoiseReduceResult {
-  /** Whether the operation succeeded */
-  success: boolean;
-  /** Path to the noise-reduced output file */
-  outputPath?: string;
-  /** Duration of the input file in seconds */
-  inputDuration?: number;
-  /** Error message on failure */
-  error?: string;
-}
-
-/**
- * Reduce audio noise in a video or audio file using FFmpeg afftdn filter.
- *
- * Supports three strength presets (low/medium/high) with optional highpass/lowpass
- * for the "high" setting. Video streams are copied without re-encoding.
- *
- * @param options - Noise reduction configuration
- * @returns Result with output path and input duration
- */
-export async function executeNoiseReduce(options: NoiseReduceOptions): Promise<NoiseReduceResult> {
-  const {
-    inputPath,
-    outputPath,
-    strength = "medium",
-    noiseFloor,
-  } = options;
-
-  if (!existsSync(inputPath)) {
-    return { success: false, error: `File not found: ${inputPath}` };
-  }
-
-  if (!commandExists("ffmpeg")) {
-    return { success: false, error: "FFmpeg not found. Install with: brew install ffmpeg (macOS) or apt install ffmpeg (Linux). Run `vibe doctor` for details." };
-  }
-
-  try {
-    const inputDuration = await getVideoDuration(inputPath);
-
-    // Map strength to noise floor dB value
-    const nf = noiseFloor ?? (strength === "low" ? -20 : strength === "high" ? -35 : -25);
-
-    // Build audio filter
-    let audioFilter = `afftdn=nf=${nf}`;
-    if (strength === "high") {
-      audioFilter = `${audioFilter},highpass=f=80,lowpass=f=12000`;
-    }
-
-    // Check if input has video stream
-    let hasVideo = false;
-    try {
-      const { stdout } = await execSafe("ffprobe", [
-        "-v", "error", "-select_streams", "v", "-show_entries", "stream=codec_type", "-of", "csv=p=0", inputPath,
-      ], { maxBuffer: 10 * 1024 * 1024 });
-      hasVideo = stdout.trim().includes("video");
-    } catch {
-      // No video stream
-    }
-
-    const args = ["-i", inputPath, "-af", audioFilter];
-    if (hasVideo) args.push("-c:v", "copy");
-    args.push(outputPath, "-y");
-    await execSafe("ffmpeg", args, { timeout: 600000, maxBuffer: 50 * 1024 * 1024 });
-
-    return {
-      success: true,
-      outputPath,
-      inputDuration,
-    };
-  } catch (error) {
-    return {
-      success: false,
-      error: `Noise reduction failed: ${error instanceof Error ? error.message : String(error)}`,
-    };
-  }
-}
-
-// ============================================================================
-// Fade
-// ============================================================================
-
-/** Options for {@link executeFade}. */
-export interface FadeOptions {
-  /** Path to the input video file */
-  videoPath: string;
-  /** Path for the output video with fade effects */
-  outputPath: string;
-  /** Fade-in duration in seconds (default: 1) */
-  fadeIn?: number;
-  /** Fade-out duration in seconds (default: 1) */
-  fadeOut?: number;
-  /** Apply fade to audio only (video copied) */
-  audioOnly?: boolean;
-  /** Apply fade to video only (audio copied) */
-  videoOnly?: boolean;
-}
-
-/** Result from {@link executeFade}. */
-export interface FadeResult {
-  /** Whether the operation succeeded */
-  success: boolean;
-  /** Path to the output video */
-  outputPath?: string;
-  /** Total duration of the source video in seconds */
-  totalDuration?: number;
-  /** Whether fade-in was applied */
-  fadeInApplied?: boolean;
-  /** Whether fade-out was applied */
-  fadeOutApplied?: boolean;
-  /** Error message on failure */
-  error?: string;
-}
-
-/**
- * Apply fade-in and/or fade-out effects to video and/or audio using FFmpeg.
- *
- * @param options - Fade configuration
- * @returns Result with output path and which fades were applied
- */
-export async function executeFade(options: FadeOptions): Promise<FadeResult> {
-  const {
-    videoPath,
-    outputPath,
-    fadeIn = 1,
-    fadeOut = 1,
-    audioOnly = false,
-    videoOnly = false,
-  } = options;
-
-  if (!existsSync(videoPath)) {
-    return { success: false, error: `Video not found: ${videoPath}` };
-  }
-
-  if (!commandExists("ffmpeg")) {
-    return { success: false, error: "FFmpeg not found. Install with: brew install ffmpeg (macOS) or apt install ffmpeg (Linux). Run `vibe doctor` for details." };
-  }
-
-  try {
-    const totalDuration = await getVideoDuration(videoPath);
-
-    const videoFilters: string[] = [];
-    const audioFilters: string[] = [];
-
-    // Video fade filters
-    if (!audioOnly) {
-      if (fadeIn > 0) {
-        videoFilters.push(`fade=t=in:st=0:d=${fadeIn}`);
-      }
-      if (fadeOut > 0) {
-        const fadeOutStart = Math.max(0, totalDuration - fadeOut);
-        videoFilters.push(`fade=t=out:st=${fadeOutStart}:d=${fadeOut}`);
-      }
-    }
-
-    // Audio fade filters
-    if (!videoOnly) {
-      if (fadeIn > 0) {
-        audioFilters.push(`afade=t=in:st=0:d=${fadeIn}`);
-      }
-      if (fadeOut > 0) {
-        const fadeOutStart = Math.max(0, totalDuration - fadeOut);
-        audioFilters.push(`afade=t=out:st=${fadeOutStart}:d=${fadeOut}`);
-      }
-    }
-
-    // Build FFmpeg command
-    const ffmpegArgs: string[] = ["-i", videoPath];
-
-    if (videoFilters.length > 0) {
-      ffmpegArgs.push("-vf", videoFilters.join(","));
-    } else if (audioOnly) {
-      ffmpegArgs.push("-c:v", "copy");
-    }
-
-    if (audioFilters.length > 0) {
-      ffmpegArgs.push("-af", audioFilters.join(","));
-    } else if (videoOnly) {
-      ffmpegArgs.push("-c:a", "copy");
-    }
-
-    ffmpegArgs.push(outputPath, "-y");
-
-    await execSafe("ffmpeg", ffmpegArgs, { timeout: 600000, maxBuffer: 50 * 1024 * 1024 });
-
-    return {
-      success: true,
-      outputPath,
-      totalDuration,
-      fadeInApplied: fadeIn > 0,
-      fadeOutApplied: fadeOut > 0,
-    };
-  } catch (error) {
-    return {
-      success: false,
-      error: `Fade failed: ${error instanceof Error ? error.message : String(error)}`,
-    };
-  }
-}
+// noise-reduce + fade moved to _shared/edit/{noise-reduce,fade}.ts (v0.69 Phase 3)
+export { executeNoiseReduce } from "./_shared/edit/noise-reduce.js";
+export type { NoiseReduceOptions, NoiseReduceResult } from "./_shared/edit/noise-reduce.js";
+export { executeFade } from "./_shared/edit/fade.js";
+export type { FadeOptions, FadeResult } from "./_shared/edit/fade.js";
 // ============================================================================
 // Translate SRT
 // ============================================================================
@@ -1547,43 +1337,9 @@ export async function executeTextOverlay(options: TextOverlayOptions): Promise<T
 // ============================================================================
 
 /** A single auto-fixable issue identified during video review. */
-export interface AutoFix {
-  /** Category of the fix */
-  type: "color_grade" | "text_overlay_adjust" | "speed_adjust" | "crop";
-  /** Human-readable description of the issue */
-  description: string;
-  /** FFmpeg filter string to apply the fix (if applicable) */
-  ffmpegFilter?: string;
-}
-
-/** Scored review for a single quality category. */
-export interface VideoReviewCategory {
-  /** Quality score from 1-10 */
-  score: number;
-  /** List of identified issues */
-  issues: string[];
-  /** Whether the issues can be auto-fixed */
-  fixable: boolean;
-  /** Suggested FFmpeg filter for fixing (color category) */
-  suggestedFilter?: string;
-  /** Improvement suggestions (text readability category) */
-  suggestions?: string[];
-}
-
-/** Complete AI video review feedback from Gemini analysis. */
-export interface VideoReviewFeedback {
-  /** Overall quality score from 1-10 */
-  overallScore: number;
-  /** Per-category quality assessments */
-  categories: {
-    pacing: VideoReviewCategory;
-    color: VideoReviewCategory;
-    textReadability: VideoReviewCategory;
-    audioVisualSync: VideoReviewCategory;
-    composition: VideoReviewCategory;
-  };
-  /** List of auto-fixable issues with FFmpeg filter suggestions */
-  autoFixable: AutoFix[];
-  /** General improvement recommendations */
-  recommendations: string[];
-}
+// video review types moved to _shared/edit/video-review.ts (v0.69 Phase 3)
+export type {
+  AutoFix,
+  VideoReviewCategory,
+  VideoReviewFeedback,
+} from "./_shared/edit/video-review.js";
